report-it-now-flask/requirements.txt
alembic==1.16.1
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pycparser==2.22
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
WTForms==3.2.1
